Feature/etapa extra average province data

+++ SimulationExercise.Core/Entities/AverageProvinceData.cs
@@ -5,14 +5,14 @@ namespace SimulationExercise.Core.Entities
     public class AverageProvinceData
     {
         public AverageProvinceData(string province,
-                                     string sensorTypeName,
-                                     double averageValue,
-                                     Unit unit,
-                                     int averageDaysOfMeasure)
+                                   string sensorTypeName,
+                                   double averageValue,
+                                   Unit unit,
+                                   int averageDaysOfMeasure)
         {
             Province = province;
             SensorTypeName = sensorTypeName;
-            AverageValue = averageValue;
+            AverageValue = (float)averageValue;
             Unit = unit;
             AverageDaysOfMeasure = averageDaysOfMeasure;
         }
